upd: remove class CValidationState from main.h
